Moved cmake code to stay with cmake
